Translated example scripts to integration tests; removed example directory
