+++ tests/unit_tests/solver/test_solver.py
